requirements.txt
anyio==4.2.0
assertpy==1.1
attrs==23.2.0
black==23.12.1
busypie==0.5.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.0
decorator==5.1.1
deprecation==2.1.0
docker==7.0.0
exceptiongroup==1.2.0
fastapi==0.108.0
google-api-core==2.15.0
google-auth==2.26.1
google-cloud-pubsub==2.19.0
google-cloud-tasks==2.15.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
isbnlib==3.10.14
isort==5.13.2
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.5
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-httpx==0.28.0
python-dateutil==2.8.2
requests==2.31.0
rfc3986==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
tenacity==8.2.3
testcontainers==3.7.1
tomli==2.0.1
typing_extensions==4.9.0
urllib3==1.26.15
uvicorn==0.25.0
websocket-client==1.7.0
wrapt==1.16.0
